Pass instance-id/hardware directly to jujud
The bootstrap-state agent currently loads its
instance ID and hardware characteristics by
fetching a URL that is written to a file on
disk. This is no longer necessary with
synchronous bootstrap.

I have changed jujud and bootstrap script
generation to pass the instance-id and hardware
characteristics directly to jujud as command
line arguments.

Fixes part of lp:1291292