environs/manual: don't require resolvable hostname
This change makes it so that the hostname provided
during manual provisioning is not necessarily
resolvable from the machine provisioning. If this
is the case, the machine will not be given a public
address.

There's another change here, which is that we no
longer perform a reverse lookup or record the
resolved addresses in state; we only store the
input address if it's an IP or if it's a resolvable
hostname. All other addresses will come from the
machine agent itself.

Fixes lp:1300264