Fix bug #129980. Refuse downgrades.
There is a race condition upgrading from 1.16 to 1.18 where the API
server and the Unit agents can upgrade before the Machine agent
finishes upgrading.

When everything is 1.16, all agents will see the need to upgrade at
the same time. If the Unit and API server finish upgrading to 1.18
first, then when the Unit agents checks again to see if it should
upgrade, it will be told that it should use the version that matches
its Machine agent (which is still 1.16). Which essentially causes
the Unit agent to want to downgrade.

However, downgrades aren't actually supported, so it just ends up
broken.