Provisioner treats no-tools as provisioning failure.
(bug #1311676) When the provisioner would fail to find tools, it would treat it
as a failure of the Provisioner and not a failure for that machine. As such,
the provisioner would fatally exit with an error, but when it woke up again it
would fail again with the same error.
Instead, treat this as a provisioning error (as though StartInstance failed), which
lets us set it in status, and lets us see that things are broken. If someone does
manage to get tools available, then they can retry-provisioning.