state: reduce default dial timeout to 30s
Previously we could have been dialling state connections
across the Internet, since the CLI would connect to Mongo
directly. Now, with the CLI going to API only, we can
safely reduce the timeout.

It is possible that the timeout will be exceeded if the
peers in an HA setup are not all ready. In this case the
dial operation will fail and the caller (state server agent)
will restart.

This change will reduce the time to failure in tests, and
allow us to capture some information about the failed test
without exceeding the Go test timeout.