Introduce lxc-use-clone, ContainerManagerConfig
A new lxc-use-clone config attribute and a new
ContainerManagerConfig provisioner API are introduced.
If lxc-use-clone is set to true, then the LXC
provisioner will be configured to use cloning
regardless of provider type. This setting is currently
immutable; additional work will be required to support
reacting to changes and reconfiguring the provisioner.

Fixes lp:1315216