Fix bug #1301353 and use /usr/lib/juju/bin/mongodb when necessary.
As the store test suite now re-uses the testing infrastructure, we want to bias
our lookup to using 'mongod' from PATH since that will have v8 in it while
/usr/lib/juju/bin/mongodb won't.

Fixed another bug while I'm here, which is that if mongo fails to start, we get a panic()
because of readLastLines trying to read on an invalid handle. There was a bug where
if cmd.Start() returned an error, we kept on trying to read data from a mongo
that never existed.